Bug 820048 - Add microsecond profiling. r=ehsan
UltraBlame original commit: c1f4bf03306a7145c7fe3eaeb345797b2fe03865

+++ tools/profiler/GeckoProfiler.h
@@ -94,7 +94,7 @@ static inline void profiler_shutdown() {};
 
 
 
-static inline void profiler_start(int aProfileEntries, int aInterval,
+static inline void profiler_start(int aProfileEntries, double aInterval,
                               const char** aFeatures, uint32_t aFeatureCount,
                               const char** aThreadNameFilters, uint32_t aFilterCount) {}
