+++ tools/profiler/platform.cpp
@@ -436,7 +436,7 @@ const char** mozilla_sampler_get_features()
 }
 
 
-void mozilla_sampler_start(int aProfileEntries, int aInterval,
+void mozilla_sampler_start(int aProfileEntries, double aInterval,
                            const char** aFeatures, uint32_t aFeatureCount,
                            const char** aThreadNameFilters, uint32_t aFilterCount)
